Extract a reusable `ArtifactCollection` implementation out of `DefaultConfiguration`.
